Add GenericPairComparator
This replaces TuplePairComparator and allows interoperability between different
TypeComparators.

TypeComparator has two new methods: extractKeys() which extracts the keys from a
value and returns them as an array of Comparable and getComparators() which
returns the TypeComparatorS for the key fields as an array.

GenericPairComparator uses these two methods two extracts keys and do
comparisons.